+++ electron_fuses/__init__.py
@@ -8,5 +8,5 @@ __license__:      str = "3-clause BSD License"
 __author_email__: str = "info@ripeda.com"
 
 from .config    import FuseConfig
-from .resources import FuseState, FuseV1Options
+from .resources import FuseState, FuseV1Options, SentinelNotFound
 from .resolve   import ResolveFramework