+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Electron Fuses
 
 ## 1.1.0
+- Add new exception for missing `SENTINEL` in binary
+  - `SentinelNotFound`
 
 ## 1.0.0
 - Initial release